getMidNode
if we have n nodes in the list, then the fast pointer will move n/2 steps forward to reach the end of linked list. During the same time, slow pointer will also move n/2 steps but without skipping any nodes in between, and hence will only make it up to the center of the linkedlist. 
so in this approach, we only run our loop for half the time... i.e n/2. this is a lot faster appraoch. 

hence, the time complexity is = O(n/2) i.e O(n). 
                  space complexity = O(2)  i.e O(1)  because we r using constant extra space here.